requirements.txt
flask==2.0.2
flask_sqlalchemy==2.5.1
python_dotenv==0.19.2